Make Configuration as a part of interface type
