+++ analyze/algorithm/BfsObjectDetector.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "IObjectDetector.h"
-#include "configuration/Configuration.hpp"
+#include "sh_mem/Message.hpp"
 #include "common/Logger.hpp"
 #include "Object.h"
 
@@ -33,11 +33,11 @@ namespace am::analyze::algorithm
 
 	protected:
 		// abstract class construction limited, aggregate common functions and interface requirement
-		BfsObjectDetector(const size_t threads, const am::configuration::Configuration &conf,
+		BfsObjectDetector(const size_t threads, const Configuration &conf,
 						  std::shared_ptr<am::common::Logger> &logger);
 
 		size_t mThreadsCount;
-		am::configuration::Configuration mConfiguration;
+		Configuration mConfiguration;
 		std::shared_ptr<am::common::Logger> mLogger;
 	};
 } // namespace am::analyze::algorithm
