+++ sh_mem/Message.hpp
@@ -26,19 +26,19 @@ struct Message
     MessageType type;
 };
 
-struct AmConfiguration
-{
-    size_t AffinityThreshold;
-    size_t MinPixelsForObject;
-    uint8_t PixelStep;
-    double CalculationTimeLimit;
-    size_t IdleTimeout;
-    double ThreadsMultiplier;
-};
+	struct Configuration
+	{
+		size_t AffinityThreshold;
+		size_t MinPixelsForObject;
+		uint8_t PixelStep;
+		double CalculationTimeLimit;
+		size_t IdleTimeout;
+		double ThreadsMultiplier;
+	};
 
 struct MessageSetConfig : public Message
 {
-    AmConfiguration configuration;
+    Configuration configuration;
 };
 
 struct MessageCompareRequest : public Message
